+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Console Logger
-- Composite Logger
 - Logger Factory
+- Composite Logger
+- Console Logger
+- Trace Logger
 
 ## [0.4.1] - 2024-10-26